requirements.txt
fpdf==1.7.2
fslpy==3.21.0
matplotlib==3.8.4
nibabel==5.2.1
nilearn==0.10.4
numpy==1.26.4
openpyxl==3.1.4
pandas==2.2.2
scikit-image==0.20.0
scikit-learn==1.5.1
scipy==1.13.1
xlsxwriter==3.2.0